Revert "Kino 22"
