+++ api/v1/urls.py
@@ -12,4 +12,5 @@ router = APIRouter(
 router.include_router(router=cinema_router, prefix="/cinema", tags=["cinema"])
 router.include_router(router=common_router, prefix="/common", tags=["common"])
 router.include_router(router=user_router, prefix="/users", tags=["users"])
+router.include_router(router=cinema_router, prefix="/cinema", tags=["cinema"])
 router.include_router(router=film_router, prefix="/films", tags=["films"])
