+++ tests/api/films/test_films_api.py
@@ -5,22 +5,17 @@ from enum import Enum
 import pytest
 from httpx import AsyncClient
 
-from apps.films.exceptions.films import FilmNotFoundException
-from apps.films.services.film_sessions import BaseFilmSessionService
 from apps.films.services.films import BaseFilmService
 from core.enums.films import AgeRatingEnum, FilmStatusEnum
-from tests.factories.film_sessions import FilmSessionFactory
-from tests.factories.films import FilmFactory
 from tests.factories.halls import HallFactory
 
 
 class TestFilmAPI:
     @staticmethod
-    def get_list_url(*args, **kwargs):
-        return "api/v1/films/" + "/".join(map(str, args))
+    def get_list_url(**kwargs):
+        return "api/v1/films/"
 
-    @staticmethod
-    def generate_fake_file(faker):
+    def generate_fake_file(self, faker):
         # Generate fake file content
         file_content = faker.binary(length=1024)  # 1KB of random bytes
         # Create a file-like object
@@ -28,20 +23,6 @@ class TestFilmAPI:
         # Set the name attribute (if required)
         file_bytes.name = "fake_image.jpg"
         return file_bytes
-
-    @staticmethod
-    async def generate_payload(faker, date_rent_start, date_rent_end):
-        hall = await HallFactory().create()
-        return {
-            "cinemahall_id": hall.id,
-            "title": faker.word(),
-            "description": faker.text(),
-            "age_rating": faker.enum(AgeRatingEnum).value,
-            "duration": faker.pyfloat(),
-            "status": faker.enum(FilmStatusEnum).value,
-            "date_rent_start": date_rent_start.isoformat(),
-            "date_rent_end": date_rent_end.isoformat(),
-        }
 
     @pytest.mark.asyncio
     async def test_create_film(
@@ -53,12 +34,19 @@ class TestFilmAPI:
         minio_cleanup,
         fake_file,
     ):
+        hall = await HallFactory().create()
         date_rent_start = datetime.now(UTC) + timedelta(4)
-        payload = await self.generate_payload(
-            faker=faker,
-            date_rent_start=date_rent_start,
-            date_rent_end=date_rent_start + timedelta(1),
-        )
+        payload = {
+            "cinemahall_id": hall.id,
+            "title": faker.word(),
+            "description": faker.text(),
+            "age_rating": faker.enum(AgeRatingEnum).value,
+            "duration": faker.pyfloat(),
+            "status": faker.enum(FilmStatusEnum).value,
+            "date_rent_start": date_rent_start.isoformat(),
+            "date_rent_end": (date_rent_start + timedelta(1)).isoformat(),
+        }
+
         files = {"poster": ("fake_image.jpg", fake_file, "image/jpeg")}
         response = await client.post(
             self.get_list_url(), data=payload, files=files
@@ -85,12 +73,18 @@ class TestFilmAPI:
         prepare_database,
         fake_file,
     ):
+        hall = await HallFactory().create()
         date_rent_start = datetime.now(UTC) - timedelta(1)
-        payload = await self.generate_payload(
-            faker=faker,
-            date_rent_start=date_rent_start,
-            date_rent_end=date_rent_start + timedelta(1),
-        )
+        payload = {
+            "cinemahall_id": hall.id,
+            "title": faker.word(),
+            "description": faker.text(),
+            "age_rating": faker.enum(AgeRatingEnum).value,
+            "duration": faker.pyfloat(),
+            "status": faker.enum(FilmStatusEnum).value,
+            "date_rent_start": date_rent_start.isoformat(),
+            "date_rent_end": (date_rent_start + timedelta(1)).isoformat(),
+        }
         files = {"poster": ("fake_image.jpg", fake_file, "image/jpeg")}
         response = await client.post(
             self.get_list_url(), data=payload, files=files
@@ -108,12 +102,18 @@ class TestFilmAPI:
         prepare_database,
         fake_file,
     ):
+        hall = await HallFactory().create()
         date_rent_start = datetime.now(UTC) - timedelta(1)
-        payload = await self.generate_payload(
-            faker=faker,
-            date_rent_start=date_rent_start,
-            date_rent_end=date_rent_start - timedelta(1),
-        )
+        payload = {
+            "cinemahall_id": hall.id,
+            "title": faker.word(),
+            "description": faker.text(),
+            "age_rating": faker.enum(AgeRatingEnum).value,
+            "duration": faker.pyfloat(),
+            "status": faker.enum(FilmStatusEnum).value,
+            "date_rent_start": date_rent_start.isoformat(),
+            "date_rent_end": (date_rent_start - timedelta(1)).isoformat(),
+        }
         files = {"poster": ("fake_image.jpg", fake_file, "image/jpeg")}
         response = await client.post(
             self.get_list_url(), data=payload, files=files
@@ -122,34 +122,3 @@ class TestFilmAPI:
         film_service = container.resolve(BaseFilmService)
         films = await film_service.get_all()
         assert not films
-
-    async def test_get_film_sessions_by_id_film(
-        self, client: AsyncClient, container, faker, prepare_database
-    ):
-        film = await FilmFactory().create()
-        film_sessions_services = container.resolve(BaseFilmSessionService)
-        amount_film_sessions = faker.pyint(min_value=1, max_value=6)
-        film_sessions = await FilmSessionFactory().create_batch(
-            amount_film_sessions
-        )
-        for session in film_sessions:
-            await film_sessions_services.update(
-                session.id, {"film_id": film.id}
-            )
-        response = await client.get(self.get_list_url(film.id, "sessions"))
-        response_json = response.json()["data"]
-        assert response.status_code == 200
-        assert amount_film_sessions == len(response_json["film_sessions"])
-        assert response_json["film_sessions"] == sorted(
-            response_json["film_sessions"], key=lambda d: d["date_time"]
-        )
-
-    async def test_get_film_sessions_by_id_not_exist_film(
-        self, client: AsyncClient, faker, prepare_database
-    ):
-        response = await client.get(
-            self.get_list_url(faker.pyint(), "sessions")
-        )
-        response_json = response.json()
-        assert response.status_code == 404
-        assert response_json["message"] == FilmNotFoundException().message
